+++ src/lang-xquery/main/uk/co/reecedunn/intellij/plugin/xquery/parser/XQueryParser.kt
@@ -1485,6 +1485,7 @@ class XQueryParser : XPathParser() {
     @Suppress("Reformat") // Kotlin formatter bug: https://youtrack.jetbrains.com/issue/KT-22518
     override fun parseExprSingleImpl(builder: PsiBuilder, parentType: IElementType?): Boolean {
         return (
+            parseWithExpr(builder) ||
             parseFLWORExpr(builder) ||
             parseQuantifiedExpr(builder) ||
             parseSwitchExpr(builder) ||